requirements.txt
# Core dependencies
flask>=2.0.0
numpy>=1.21.0
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
open_clip_torch>=2.0.0
faiss-gpu>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Image processing
opencv-python>=4.5.0
imageio>=2.9.0

# Utilities
tqdm>=4.62.0
requests>=2.25.0
python-dotenv>=0.19.0

# Optional: for GPU monitoring
nvidia-ml-py3>=7.352.0
psutil>=5.8.0 